pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "minecraft-diagnostic-mcp"
version = "0.1.0"
description = "Model Context Protocol server for Minecraft diagnostics across backup, local runtime, and Docker runtime environments."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "mcp[cli]>=1.2.1",
  "PyYAML>=6.0.2",
]

[project.scripts]
minecraft-diagnostic-mcp = "minecraft_diagnostic_mcp.__main__:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
